Fixed bug causing strange throughput while generating 8bit array in test. Also, some variables were renamed to avoid such errors in future.
Explanation: count parameter in _array function is count of 8/16/32/64 bit numbers, and returned value means the same. But the test was always giving the same count to 8 and 64 bit generator, so the 8bit was in fact generating only 1/8 of desired bytes and because the expected value was amount of 64bit blocks, it didn't threw an error.
Correct usage for filling array of 64bit blocks is rdrand_get_uint8_array_retry((uint64_t*)&buf[i*chunk], chunk*8, 1)/8.
In case of opposite situation (filling array of 8bit blocks with 64bit array function), the chunk size has to be divided by 8.